OMAP: HDMI: Kernel-side workaround for bug 6347047.
Summary of this issue is that audio and video on HDMI in OMAP know
little about each other, and their lack of coordination is causing
kernel panics.

More Detail:  When the HDMI cable is unplugged from a device, the
video layer of the system springs into action and starts to shut down
the HDMI unit (as it should).  While it is responsible for shutting
down all things video, it did not handle shutting down the audio path.
Instead it sent a signal to the application level via a switch device
which is supposed to cause the application level to shut down audio
(if it had been started).  If the application level is slow, or not
paying attention, or buggy (or anything else) and the audio does not
actually get shut down promptly, bad things start to happen.  The
video shutdown code blocks for 10 seconds waiting for the audio to
shut down, and after those 10 seconds have timed out, it goes ahead
and shuts down the unit anyway.  Later on, when the app layer catches
up and attempts to shut down audio, it touches de-clocked, de-powered
registers and triggers a kernel panic.

I would really like to see TI attempt to rework a lot of their HDMI
code to treat the HDMI interconnect as a resource which is shared
between audio, video, CEC, and anything else it carries, instead of
just treating it like a video only piece of HW and having almost no
coordination between audio and video.  In the meantime, we have this
workaround.  Basically, when the cable is unplugged, the shutdown
state machine calls directly into the audio driver (if it is compiled
in) and forces it into a shutdown state out from underneath the
application level.  The app will need to close and reopen its driver
handle in order to recover, and the open will not succeed until the
cable is plugged in to a sync which supports audio and has an active
video mode set up and running.  This makes it so that the kenel level
will not crash no matter how ill behaved the application level is.

Change-Id: I3375a292e443f96685126eec8585befcb037e178
Signed-off-by: John Grossman <johngro@google.com>
(cherry picked from commit fab34b3e1d28304a0f47a9624588668fc58c64f3)