steelhead_avr: Add wipe request support during early boot
The current method to request a factory reset/wipe in
the bootloader has too small of a window for the user
to trigger it easily/reliably.

By reproducing the same logic in the kernel, the user
has a larger window of time to initiate the wipe.
The trigger is still a 10 second hold of the mute key,
but now the window is between the time the kernel boots
and when userland makes the first ioctl to change led
mode to host control, which is about 16 seconds into boot.
So if they start the press within first 5 seconds or so of
boot, it'll work (whereas before they had about 1 second
or less of a window where the bootloader was checking).

Bug: 6438620

Change-Id: I2f81d8d2ff4fa7557205de262652cb38d792a92a
Signed-off-by: Mike J. Chen <mjchen@google.com>
(cherry picked from commit b3303549e61df559632ee4a141b0ec1434cbe40d)