NFC: Prevent a lock up condition in the NFC stack.
Add some protection so that a glitch on the IRQ GPIO cannot cause a
lockup in the NFC read function.  The trouble comes when using the
current state of the GPIO as the wake condition for the wake lock
used in the read function.  The sequence of events leading up to the
lock-up is as follows.

1) Client code calls read.  Read un-masks the IRQ and enters the
   wait_event_interruptable macro.
2) WEI checks the condition (reads the GPIO) and finds to to be low.
3) WEI adds the thread to the waitqueue and sets the status to
   TASK_INTERRUPABLE
4) WEI checks the condition (reads the GPIO) and finds to to be low
   again.  It then yeilds waiting to be woken up again.
5) GPIO goes high and IRQ fires.  IRQ handler wakes up, masks the
   interrupt and then wakes up the members of the wait queue, then
   exits.
6) Reader thread is scheduled, it adds it self back to the wait queue
   and then checks the GPIO.  If the GPIO had glitched, it sees the
   GPIO as being low and then simply yields waiting to be woken up
   again.
7) GPIO goes high for good.  IRQ is masked, so nothing happens.  At
   this point, the reader thread is de-scheduled and there is nothing
   left to ever wake it back up again.

Moral of the story here is that the conditions for waitqueues really
cannot be transient when used in this single waker pattern.  When the
IRQ handler fires, it needs to set a sticky condition because if the
reader thread does not wake up because of a transient condition, the
system will be stuck forever.

This CL tweaks things to use an internal state variable (which is
sticky) to wake up the reader.  The reader then checks the GPIO, and
if its low, it will report a spurious interrupt, re-enable the IRQ and
then enter the waitqueue again.  If the GPIO ever goes high and stays
high, the system will unblock and read as it is supposed to.

Change-Id: I14007d7069dbeae8194bc2723dc99e51a2eb3701
Signed-off-by: John Grossman <johngro@google.com>
(cherry picked from commit bcd1167a114fa20dba801907eec1fb23eb9657d3)