OMAP4: HDMI: Don't allow set mode if edid isn't set
There's a race condition possible when hpd is bouncing
where the hwc is trying to set the mode from a previous
hpd->1 case but the hdp state machine is running to handle
a new hpd->1 case (with an hpd->0 in between).  Since
setting the mode disables the driver, the mode set could
disrupt the edid read and cause it to fail/abort.

Adding a check at the beginning of set mode for edid read
isn't fool proof but reduces the race possibilty considerably.
It's hard to fix it completely because the set mode is
a sequence of non-atomic operations.

Change-Id: Iec3c091098148f2e2cc8d01917e2bb37e56e852d
Signed-off-by: Mike J. Chen <mjchen@google.com>
(cherry picked from commit fee8b90e629e40324d93a1b8b2c6a2f69127e852)