OMAP4: HDMI: Rewrite initial hpd handling
A major rewrite of the way the hdmi driver works after
hpd is detected.

Separate the notion of powering up hdmi for the
purpose of reading edid and turning on the display.
The latter is done in set mode, after edid is read
and the modes supported by the sink are known.

Enable a video mode in the kernel driver once again,
but this time take the preferred mode from the board
file and validate it against the modes supported by
the sink.  Fallback to VGA if not supported.  This
helps eliminate (in my testing with my limited set
of AVRs and monitors) hpd bounces by sinks that aren't
happy if we don't enable video soon after edid exchange
(they either drop hpd a second or so after the edid
exchange without sending video, or much later when
we do set a video mode many seconds after edid exchange).

Change-Id: I9f3761bca30e8687e58f0ab625c622b4263f338b
(cherry picked from commit 73982d44f792e9cc5c0c0d8ffab810d726abd6e8)