OMAP4: HDMI: Improve native and initial video mode logic
Don't assume VGA is a safe default video mode.
When the preferred mode isn't supported by the monitor,
we used to default to VGA, but turns out some monitors
don't even support that (they might support D1 instead,
which the HDMI spec allows).  So, change the logic
so that if the preferred mode isn't supported, we choose
a mode that is supported according to the EDID parsing.
It will either be the first supported mode in the EDID,
or the preferred mode the monitor specified (if supported
by the HW, e.g. we don't choose interlaced modes even if
the monitor says it supports it).

Use SVD native resolution when better than DTD preferred.
The SVD native resolution is a newer EDID field that can
provide better information about the native resolution
of sink.  Use it if the resolution is better than the
one in the DTD first entry.  We may always want to use
it if it's available instead of the DTD first entry, but
this is more conservative for now.

Change-Id: I7c81b7d530e8a2b5b80a8afd4e30b587612ccc0e
Signed-off-by: Mike J. Chen <mjchen@google.com>
(cherry picked from commit dae13c454edb81a32d45c076ae70e293ac13607a)