OMAP4: HDCP: Check hpd at end of hdcp_load_keys_cb
Boot time hpd check used to be in hdmi_init, but that's
too early because hdcp keys are loaded later and the hpd
handling wants to start hdcp authentication and that would
fail and never try again.  So we removed the hpd check
in hdmi_init and put it in the hdcp_load_keys_cb().  The
previous implementation only called the hpd handler if
the keys loaded successfully but it should be called even
if the keys failed to load (which can happen in a factory
setting before keys have been provisioned).

Change-Id: I70c4e830b5b4f495c8e59884882d140010e289e2
Signed-off-by: Mike J. Chen <mjchen@google.com>
(cherry picked from commit 57b3811dc5fbb550e5fe69291403b36751e362ca)