OMAP4: HDCP: Only do AV_MUTE during reauthentication
If a HDMI sink does not support HDCP, the driver will keep
trying, but will constantly blink the video because the
authentication code always did an AV_MUTE operation.

Change the logic to only do the AV_MUTE on reauthentication
after a previously successful authentication.  The AV_MUTE
is mainly to synchronize the video when transitioning from
non-encrypted to encrypted (and vice versa) but since no
transition ever happens if HDCP never authenticated during
a HPD session, it's not needed if we've never succeeded in
authenticating.

This should fix the blinking problem for the case of HDMI
sinks that don't support HDCP.

Change-Id: I20614571c0f585b8383ed47ef6bb9654162cbfca
Signed-off-by: Mike J. Chen <mjchen@google.com>
(cherry picked from commit 584b850a209f4f915da5157101758addde4fc69c)